Notify the other RootSync when management conflict error is detected
If management conflict error is detected, both remediator will report
the error. Originally, we thought it is okay to leverage the
reconciler's remediator to update the RootSync status.

However, if conflict is detected in a remediator that syncs from many
resources, the corresponding remediator can surface the error and the
kpt_applier will retry these resources, which takes long time. The other
RootSync will be in a successful state for the retry time.

To reflect the conflict earlier, it's better to notify both RootSync
objects. This commit also dedups the conflict error surfaced by the
current remediator and reported from the other conflict remediator.

It also removes all conflict errors for objects with the same GVK when a
watch for the GVK is stopped because the objects are no longer managed
by the reconciler.

b/213343423

Change-Id: I8e98b5556e0eeb611481c4e33087e01dcaf2d4bd