Add a field to control the apply library to embed apply status
This is part of the effort to solving the scalability issue in
b/222554711. Since embedding the apply status increases
the ResourceGroup object size, we should support users
to opt out.

With this change, users can set the field `.spec.override.statusMode`
to `disabled`. Then the kpt applier won't update the ResourceGroup
status field.

Bug: b/222595604

Change-Id: I8ef23b7dc5b73db80ecfa1af175ac3988ffb7d41