Stop watching ConfigMaps to reduce the reconciler-manager memory usage
The reposync-controller in the reconciler-manager only cares about ConfigMaps
in the config-management-system namespace. Only mutations in the namespace can
trigger a reconciliation of the RepoSync CR. The c-m-s namespace is kind of
locked by RBAC, which means it requires privilege to mutate resources in the
namespace. With such privilege, it seems like a user/controller could mutate
the reconciler manager and/or operator anyways.

Hence, this CL drops the ConfigMaps watch. As a follow-up, instead of
passing environment variables through the ConfigMaps, the reposync-controller
can just modify the manifest of the reconciler to specify the environment
variables directly.

b/223203740

Change-Id: I8e2deaa5bf5ae00bad7e3daad6ce16f5bcd9d18e