Stop watching ConfigMaps for RootSync object
The rootsync-controller also watches ConfigMaps using the .Owns() call,
which still caches all ConfigMaps in memory. An alternative is to let
the rootsync-controller only cache resources in the c-m-s namespace,
but it requires changing the controller manager, which is shared
between rootsync-controller and reposync-controller.

To see the memory usage enhancement in 1.11.0, we have stop watching
ConfigMaps for RootSync object as well.

b/223248327

Change-Id: I11f9d3a015c3045333cb24a1db01b8780b1ec39d