Inject fleet workload identity credentials to askpass sidecar
To support authenticating with fleet workload identity, the
reconciler-manager needs to pass the crendentials file to the askpass
sidecar. The parameters in the credentials file include GSA email and
fleet membership details.

This commit extractst the membership details using the membership API
and injects the credentials file to the askpass sidecar.

b/224569038
b/224569016

Change-Id: Ibd11bf4ad3e68f045b1f06629f928a46f3d8f19e