Clear the fleet membership cache when Membership CR is not found
When a cluster is unregistered from a fleet, hub deletes the Membership
CR and the CRD. There is a short period that the Membership CRD has not
been deleted yet. Hence, when we clear the membership cache for an
unregister event, we should check the existence of Membership CR instead
of CRD.

With the change, we can move the `fleetMembershipCRDExists` function
from reconciler_base.go to cmd/reconciler-manager/main.go. The function
will be called only once in the startup time.

Change-Id: I2648ee0833992d9fde3341570aedfec4c8ed0c98