Reset the initial test repo url
When the e2e tests are running in the shared test environment (GKE
clusters), the R*Sync repo url needs to be reset to the original value.
Some tests are running with other git providers (bitbucket, gitlab), so
we can't use a fixed value. The value can be extracted from the git
provider.

b/227445467
b/227302564

Change-Id: Idddf170b0168cb3bdcea1303bbd7453920bc9f91