Move NAMESPACE_NAME env var to the config maps and correct the value
Now setting this env var in reconciler and hydration controller
config maps.

Removing the variable from deployment template.

Scope is used as the value of variable, which will be either
:root for rootsyncs, <namespace> for namespace reconcilers.

This variable is mainly used for registering k8s resource labels,
so is not stressing the cardinality under custom metric namespace.

b/220343966

Change-Id: Ie2ed4e9f10b95588fdc4da4dccd8cbc6e1797628