Use a new service account to run the Prow jobs
Previously, the config-sync-ci Prow jobs use service account
`prober-runner@nomos-e2e-test1.iam.gserviceaccount.com` to run the
end-to-end test. The service account is in the nomos-e2e-test1 project,
which is in the nomos-e2e-joonix.net domain. To avoid allowing an
external domain, we should use a service account in google.com.

b/228371689

Change-Id: I453fc25dbb43c64a759099844047fc485de9b9b7