Allow overriding reconciler resources on Autopilot clusters
Autopilot enforces some rules on the workload resources, e.g. increments of
250mCPU, keeps the mem/CPU ratio of 1-6.5 GiB / CPU and etc. To allow customers
to increase memory limit when OOMing and to avoid reconciler-manager fighting
with Autopilot, we decided to make ConfigSync aware of the Autopilot policies.
When .spec.override is specified on an Autopilot cluster, validate if the sum
of the container resource requests of the reconciler Pod meets the Autopilot requirements.

See go/cs-autopilot for more details.

b/222723468

Change-Id: I242274300d62bce98d102022d06d91e384ef95bd