Add OciStatus field to the R*Sync APIs
The source/rendering/sync fields in `.status` reference the git configs.
To support OCI source type, this commit adds the OciStatus field. It
should be mutually exclusive to the GitStatus field.

It also updates the description of the `commit` fields in the API. For
backward compatibility, we didn't rename it to `hash`. Instead, we reuse
the same field to represent both the git commit hash and OCI image
digest.

b/230148287

Change-Id: If4b8cca8aee48c14f055bd262740ecff17d04b1e