Set the sourceType to git if not specified
The R*Sync objects can be applied to the cluster in the following ways:
- kubectl apply
- managed by Config Sync

If it is created via `kubectl apply` and sourceType is not specified,
.spec.sourceType will be set to `git` by default with
`+kubebuilder:default:=git`.

If it is previously managed by Config Sync, and sourceType is not specified,
instead of returning a validation error, this commit sets the sourceType to
`git` for backward compatibility concern.

It also adds validation for redundant source specifications. If both .spec.git
and .spec.oci are specified, it returns an error.

b/233681357

Change-Id: I5a6c1e258f5ad4ee5399c828d05cbf5e0464f3a1