fix: check for sync errors before assigning ref
The errorSourceRef field was always being set during the SetSyncing
calls, leading to a transient state where the errorSourceRef was set
even if an actual error was not present at the location. This change
checks to ensure that an error actually exists at the location before
assigning a value to the errorSourceRef field.

b/234663605

Change-Id: Ic1cff8bf68c0550e5990bf824bc1ffe2a00a4e5e