preload images into kind cluster
There is a test failure on the new OSS prow cluster where containerd on
the local kind cluster is unable to reach the local kind registry. The
root cause has been difficult to diagnose, so this is intended as a
workaround by preloading the local images into the kind cluster.

Change-Id: I1646f3edee6c308800efd7cd34cc71c45175c034